Add support for if/else statements and ternary expressions
Only rvalue ternary expressions are supported (I don't know if C has lvalue ternary expressions, or if only C++ does. I suspect the latter currently).
I also handle the edge case as described in: https://stackoverflow.com/questions/69311287/why-is-using-assignments-in-ternary-operator-generating-an-error-without-parenth
To elaborate:
```c
int main() {
    int a = 2;
    int b = 1;
    a > b ? a = 1 : a = 0;
    return a;
}
```
is **not** a valid C program (because it gets parsed as `(a > b ? a = 1 : a) = 0;`).
But this **is** a valid C program:
```c
int main() {
    int a = 2;
    int b = 1;
    a > b ? a = 1 : (a = 0);
    return a;
}
```
For reference, **both** of these are valid C++ programs.
This compiler properly rejects the first program for lvalue reasons, but it properly parses the second program.
This follows the behavior of gcc and the C standard with respect to ternary expressions.

Signed-off-by: LincolnSand <sandlincoln@gmail.com>