Add support for type validation and implicit type conversions
The compiler now supports integer promotion and other implicit type conversions for built-in operators using primitive types.
It also does validation that the operators are valid or supported for the types of the operands as well as checking convertibility of types.
It is written in a generic way, so it should work without any changes at all when `float`, `short`, and other similar primitives are added (the compiler currently only has `int`, `double`, and `char`).
NOTE: There is no `bool` type, so `int` is used instead in any situations where a boolean type would ordinarily be expected.
TODO: add support for type conversion for expression parameters passed in function calls.

Signed-off-by: LincolnSand <sandlincoln@gmail.com>